Changes to cachematrix as per assignment
The changes made in this commit reflect the coding of both functions,
makeCacheMatrix and cacheSolve.

makeCacheMatrix is a #This function creates an object of special Matrix
for Assigment #2  that can cache its inverse matrix

While the cacheSolve function checks if an inversed matrix cached
already for the matrix in makeCacheMatrix. If inversed matrix is in
cache, then returns it to the calling function, with a message.
Otherwise, it creates the Inverse of the supplied matrix and returns the
inverse to the makeCacheMatrix function.